requirements.txt
asgiref==3.7.2
Django==5.0
django-types==0.19.1
mysqlclient==2.2.1
nulltype==2.3.1
plaid-python==19.0.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
types-psycopg2==2.9.21.20240218
typing_extensions==4.8.0
urllib3==2.2.1
